[IMP] apply messages logic for bhdleon messages
